sql/catalog/descs: break up monolithic collection.go
The file was so big and unorderly it was hard to see what was happening.
This commit is purely movement, no new code.

Release note: None
