+++ pkg/sql/catalog/descs/BUILD.bazel
@@ -4,11 +4,18 @@ go_library(
     name = "descs",
     srcs = [
         "collection.go",
+        "database.go",
+        "descriptor.go",
         "dist_sql_type_resolver.go",
+        "hydrate.go",
         "kv_descriptors.go",
         "leased_descriptors.go",
+        "object.go",
+        "schema.go",
         "synthetic_descriptors.go",
+        "table.go",
         "txn.go",
+        "type.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/catalog/descs",
     visibility = ["//visibility:public"],
