+++ pkg/sql/catalog/descs/collection.go
@@ -15,8 +15,6 @@ package descs
 import (
 	"bytes"
 	"context"
-	"fmt"
-	"strings"
 
 	"github.com/cockroachdb/cockroach/pkg/keys"
 	"github.com/cockroachdb/cockroach/pkg/kv"
@@ -25,23 +23,15 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catalogkeys"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catalogkv"
-	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catconstants"
-	"github.com/cockroachdb/cockroach/pkg/sql/catalog/dbdesc"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/hydratedtables"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/lease"
-	"github.com/cockroachdb/cockroach/pkg/sql/catalog/schemadesc"
-	"github.com/cockroachdb/cockroach/pkg/sql/catalog/systemschema"
-	"github.com/cockroachdb/cockroach/pkg/sql/catalog/tabledesc"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/typedesc"
-	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
-	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlerrors"
 	"github.com/cockroachdb/cockroach/pkg/util/encoding"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
-	"github.com/cockroachdb/cockroach/pkg/util/protoutil"
 	"github.com/cockroachdb/errors"
 )
 
@@ -139,139 +129,6 @@ func (tc *Collection) MaybeUpdateDeadline(ctx context.Context, txn *kv.Txn) (err
 	return tc.leased.maybeUpdateDeadline(ctx, txn)
 }
 
-// GetMutableDatabaseByName returns a mutable database descriptor with
-// properties according to the provided lookup flags. RequireMutable is ignored.
-func (tc *Collection) GetMutableDatabaseByName(
-	ctx context.Context, txn *kv.Txn, name string, flags tree.DatabaseLookupFlags,
-) (*dbdesc.Mutable, error) {
-	flags.RequireMutable = true
-	desc, err := tc.getDatabaseByName(ctx, txn, name, flags)
-	if err != nil || desc == nil {
-		return nil, err
-	}
-	return desc.(*dbdesc.Mutable), nil
-}
-
-// GetImmutableDatabaseByName returns an immutable database descriptor with
-// properties according to the provided lookup flags. RequireMutable is ignored.
-func (tc *Collection) GetImmutableDatabaseByName(
-	ctx context.Context, txn *kv.Txn, name string, flags tree.DatabaseLookupFlags,
-) (catalog.DatabaseDescriptor, error) {
-	flags.RequireMutable = false
-	return tc.getDatabaseByName(ctx, txn, name, flags)
-}
-
-// GetDatabaseDesc implements the Accessor interface.
-//
-// TODO(ajwerner): This exists to support the SchemaResolver interface and
-// should be removed or adjusted.
-func (tc *Collection) GetDatabaseDesc(
-	ctx context.Context, txn *kv.Txn, name string, flags tree.DatabaseLookupFlags,
-) (desc catalog.DatabaseDescriptor, err error) {
-	return tc.getDatabaseByName(ctx, txn, name, flags)
-}
-
-// getDatabaseByName returns a database descriptor with properties according to
-// the provided lookup flags.
-func (tc *Collection) getDatabaseByName(
-	ctx context.Context, txn *kv.Txn, name string, flags tree.DatabaseLookupFlags,
-) (catalog.DatabaseDescriptor, error) {
-	if name == systemschema.SystemDatabaseName {
-		// The system database descriptor should never actually be mutated, which is
-		// why we return the same hard-coded descriptor every time. It's assumed
-		// that callers of this method will check the privileges on the descriptor
-		// (like any other database) and return an error.
-		if flags.RequireMutable {
-			proto := systemschema.MakeSystemDatabaseDesc().DatabaseDesc()
-			return dbdesc.NewBuilder(proto).BuildExistingMutableDatabase(), nil
-		}
-		return systemschema.MakeSystemDatabaseDesc(), nil
-	}
-
-	getDatabaseByName := func() (found bool, _ catalog.Descriptor, err error) {
-		if found, refuseFurtherLookup, desc, err := tc.getSyntheticOrUncommittedDescriptor(
-			keys.RootNamespaceID, keys.RootNamespaceID, name, flags.RequireMutable,
-		); err != nil || refuseFurtherLookup {
-			return false, nil, err
-		} else if found {
-			log.VEventf(ctx, 2, "found uncommitted descriptor %d", desc.GetID())
-			return true, desc, nil
-		}
-
-		if flags.AvoidCached || flags.RequireMutable || lease.TestingTableLeasesAreDisabled() {
-			return tc.kv.getByName(
-				ctx, txn, keys.RootNamespaceID, keys.RootNamespaceID, name, flags.RequireMutable,
-			)
-		}
-
-		desc, shouldReadFromStore, err := tc.leased.getByName(
-			ctx, txn, keys.RootNamespaceID, keys.RootNamespaceID, name)
-		if err != nil {
-			return false, nil, err
-		}
-		if shouldReadFromStore {
-			return tc.kv.getByName(
-				ctx, txn, keys.RootNamespaceID, keys.RootNamespaceID, name, flags.RequireMutable,
-			)
-		}
-		return true, desc, nil
-	}
-
-	found, desc, err := getDatabaseByName()
-	if err != nil {
-		return nil, err
-	} else if !found {
-		if flags.Required {
-			return nil, sqlerrors.NewUndefinedDatabaseError(name)
-		}
-		return nil, nil
-	}
-	db, ok := desc.(catalog.DatabaseDescriptor)
-	if !ok {
-		if flags.Required {
-			return nil, sqlerrors.NewUndefinedDatabaseError(name)
-		}
-		return nil, nil
-	}
-	if dropped, err := filterDescriptorState(db, flags.Required, flags); err != nil || dropped {
-		return nil, err
-	}
-	return db, nil
-}
-
-// GetObjectDesc looks up an object by name and returns both its
-// descriptor and that of its parent database. If the object is not
-// found and flags.required is true, an error is returned, otherwise
-// a nil reference is returned.
-//
-// TODO(ajwerner): clarify the purpose of the transaction here. It's used in
-// some cases for some lookups but not in others. For example, if a mutable
-// descriptor is requested, it will be utilized however if an immutable
-// descriptor is requested then it will only be used for its timestamp and to
-// set the deadline.
-func (tc *Collection) GetObjectDesc(
-	ctx context.Context, txn *kv.Txn, db, schema, object string, flags tree.ObjectLookupFlags,
-) (desc catalog.Descriptor, err error) {
-	if isVirtual, desc, err := tc.maybeGetVirtualObjectDesc(
-		schema, object, flags, db,
-	); isVirtual || err != nil {
-		return desc, err
-	}
-	// Fall back to physical descriptor access.
-	switch flags.DesiredObjectKind {
-	case tree.TypeObject:
-		typeName := tree.MakeNewQualifiedTypeName(db, schema, object)
-		_, desc, err := tc.getTypeByName(ctx, txn, &typeName, flags)
-		return desc, err
-	case tree.TableObject:
-		tableName := tree.MakeTableNameWithSchema(tree.Name(db), tree.Name(schema), tree.Name(object))
-		_, desc, err := tc.getTableByName(ctx, txn, &tableName, flags)
-		return desc, err
-	default:
-		return nil, errors.AssertionFailedf("unknown desired object kind %d", flags.DesiredObjectKind)
-	}
-}
-
 func (tc *Collection) maybeGetVirtualObjectDesc(
 	schema string, object string, flags tree.ObjectLookupFlags, db string,
 ) (isVirtual bool, _ catalog.Descriptor, _ error) {
@@ -299,827 +156,10 @@ func (tc *Collection) maybeGetVirtualObjectDesc(
 	return true, desc.Desc(), nil
 }
 
-func (tc *Collection) getObjectByName(
-	ctx context.Context,
-	txn *kv.Txn,
-	catalogName, schemaName, objectName string,
-	flags tree.ObjectLookupFlags,
-) (found bool, _ catalog.Descriptor, err error) {
-
-	// If we're reading the object descriptor from the store,
-	// we should read its parents from the store too to ensure
-	// that subsequent name resolution finds the latest name
-	// in the face of a concurrent rename.
-	avoidCachedForParent := flags.AvoidCached || flags.RequireMutable
-	// Resolve the database.
-	db, err := tc.GetImmutableDatabaseByName(ctx, txn, catalogName,
-		tree.DatabaseLookupFlags{
-			Required:       flags.Required,
-			AvoidCached:    avoidCachedForParent,
-			IncludeDropped: flags.IncludeDropped,
-			IncludeOffline: flags.IncludeOffline,
-		})
-	if err != nil || db == nil {
-		return false, nil, err
-	}
-	dbID := db.GetID()
-
-	// Resolve the schema.
-	scDesc, err := tc.GetImmutableSchemaByName(ctx, txn, dbID, schemaName,
-		tree.SchemaLookupFlags{
-			Required:       flags.Required,
-			AvoidCached:    avoidCachedForParent,
-			IncludeDropped: flags.IncludeDropped,
-			IncludeOffline: flags.IncludeOffline,
-		})
-	if err != nil || scDesc == nil {
-		return false, nil, err
-	}
-	schemaID := scDesc.GetID()
-
-	if isVirtual, desc, err := tc.maybeGetVirtualObjectDesc(
-		scDesc.GetName(), objectName, flags, db.GetName(),
-	); isVirtual || err != nil {
-		return isVirtual, desc, err
-	}
-
-	if found, refuseFurtherLookup, desc, err := tc.getSyntheticOrUncommittedDescriptor(
-		dbID, schemaID, objectName, flags.RequireMutable,
-	); err != nil || refuseFurtherLookup {
-		return false, nil, err
-	} else if found {
-		log.VEventf(ctx, 2, "found uncommitted descriptor %d", desc.GetID())
-		return true, desc, nil
-	}
-
-	// TODO(vivek): Ideally we'd avoid caching for only the
-	// system.descriptor and system.lease tables, because they are
-	// used for acquiring leases, creating a chicken&egg problem.
-	// But doing so turned problematic and the tests pass only by also
-	// disabling caching of system.eventlog, system.rangelog, and
-	// system.users. For now we're sticking to disabling caching of
-	// all system descriptors except role_members, role_options, and users
-	// (i.e., the ones used during authn/authz flows).
-	// TODO (lucy): Reevaluate the above. We have many more system tables now and
-	// should be able to lease most of them.
-	isAllowedSystemTable := objectName == systemschema.RoleMembersTable.GetName() ||
-		objectName == systemschema.RoleOptionsTable.GetName() ||
-		objectName == systemschema.UsersTable.GetName() ||
-		objectName == systemschema.JobsTable.GetName() ||
-		objectName == systemschema.EventLogTable.GetName()
-	avoidCache := flags.AvoidCached || flags.RequireMutable || lease.TestingTableLeasesAreDisabled() ||
-		(catalogName == systemschema.SystemDatabaseName && !isAllowedSystemTable)
-	if avoidCache {
-		return tc.kv.getByName(
-			ctx, txn, dbID, schemaID, objectName, flags.RequireMutable,
-		)
-	}
-
-	desc, shouldReadFromStore, err := tc.leased.getByName(
-		ctx, txn, dbID, schemaID, objectName)
-	if err != nil {
-		return false, nil, err
-	}
-	if shouldReadFromStore {
-		return tc.kv.getByName(
-			ctx, txn, dbID, schemaID, objectName, flags.RequireMutable,
-		)
-	}
-	return true, desc, nil
-}
-
-// GetMutableTableByName returns a mutable table descriptor with properties
-// according to the provided lookup flags. RequireMutable is ignored.
-func (tc *Collection) GetMutableTableByName(
-	ctx context.Context, txn *kv.Txn, name tree.ObjectName, flags tree.ObjectLookupFlags,
-) (found bool, _ *tabledesc.Mutable, _ error) {
-	flags.RequireMutable = true
-	found, desc, err := tc.getTableByName(ctx, txn, name, flags)
-	if err != nil || !found {
-		return false, nil, err
-	}
-	return true, desc.(*tabledesc.Mutable), nil
-}
-
-// GetImmutableTableByName returns a mutable table descriptor with properties
-// according to the provided lookup flags. RequireMutable is ignored.
-func (tc *Collection) GetImmutableTableByName(
-	ctx context.Context, txn *kv.Txn, name tree.ObjectName, flags tree.ObjectLookupFlags,
-) (found bool, _ catalog.TableDescriptor, _ error) {
-	flags.RequireMutable = false
-	found, desc, err := tc.getTableByName(ctx, txn, name, flags)
-	if err != nil || !found {
-		return false, nil, err
-	}
-	return true, desc, nil
-}
-
-// getTableByName returns a table descriptor with properties according to the
-// provided lookup flags.
-func (tc *Collection) getTableByName(
-	ctx context.Context, txn *kv.Txn, name tree.ObjectName, flags tree.ObjectLookupFlags,
-) (found bool, _ catalog.TableDescriptor, err error) {
-	found, desc, err := tc.getObjectByName(
-		ctx, txn, name.Catalog(), name.Schema(), name.Object(), flags)
-	if err != nil {
-		return false, nil, err
-	} else if !found {
-		if flags.Required {
-			return false, nil, sqlerrors.NewUndefinedRelationError(name)
-		}
-		return false, nil, nil
-	}
-	table, ok := desc.(catalog.TableDescriptor)
-	if !ok {
-		if flags.Required {
-			return false, nil, sqlerrors.NewUndefinedRelationError(name)
-		}
-		return false, nil, nil
-	}
-	if table.Adding() && table.IsUncommittedVersion() &&
-		(flags.RequireMutable || flags.CommonLookupFlags.AvoidCached) {
-		// Special case: We always return tables in the adding state if they were
-		// created in the same transaction and a descriptor (effectively) read in
-		// the same transaction is requested. What this basically amounts to is
-		// resolving adding descriptors only for DDLs (etc.).
-		// TODO (lucy): I'm not sure where this logic should live. We could add an
-		// IncludeAdding flag and pull the special case handling up into the
-		// callers. Figure that out after we clean up the name resolution layers
-		// and it becomes more Clear what the callers should be.
-		return true, table, nil
-	}
-	if dropped, err := filterDescriptorState(
-		table, flags.Required, flags.CommonLookupFlags,
-	); err != nil || dropped {
-		return false, nil, err
-	}
-	hydrated, err := tc.hydrateTypesInTableDesc(ctx, txn, table)
-	if err != nil {
-		return false, nil, err
-	}
-
-	return true, hydrated, nil
-}
-
-// GetMutableTypeByName returns a mutable type descriptor with properties
-// according to the provided lookup flags. RequireMutable is ignored.
-func (tc *Collection) GetMutableTypeByName(
-	ctx context.Context, txn *kv.Txn, name tree.ObjectName, flags tree.ObjectLookupFlags,
-) (found bool, _ *typedesc.Mutable, _ error) {
-	flags.RequireMutable = true
-	found, desc, err := tc.getTypeByName(ctx, txn, name, flags)
-	if err != nil || !found {
-		return false, nil, err
-	}
-	return true, desc.(*typedesc.Mutable), nil
-}
-
-// GetImmutableTypeByName returns a mutable type descriptor with properties
-// according to the provided lookup flags. RequireMutable is ignored.
-func (tc *Collection) GetImmutableTypeByName(
-	ctx context.Context, txn *kv.Txn, name tree.ObjectName, flags tree.ObjectLookupFlags,
-) (found bool, _ catalog.TypeDescriptor, _ error) {
-	flags.RequireMutable = false
-	return tc.getTypeByName(ctx, txn, name, flags)
-}
-
-// getTypeByName returns a type descriptor with properties according to the
-// provided lookup flags.
-func (tc *Collection) getTypeByName(
-	ctx context.Context, txn *kv.Txn, name tree.ObjectName, flags tree.ObjectLookupFlags,
-) (found bool, _ catalog.TypeDescriptor, err error) {
-	found, desc, err := tc.getObjectByName(
-		ctx, txn, name.Catalog(), name.Schema(), name.Object(), flags)
-	if err != nil {
-		return false, nil, err
-	} else if !found {
-		if flags.Required {
-			return false, nil, sqlerrors.NewUndefinedTypeError(name)
-		}
-		return false, nil, nil
-	}
-	typ, ok := desc.(catalog.TypeDescriptor)
-	if !ok {
-		if flags.Required {
-			return false, nil, sqlerrors.NewUndefinedTypeError(name)
-		}
-		return false, nil, nil
-	}
-	if dropped, err := filterDescriptorState(typ, flags.Required, flags.CommonLookupFlags); err != nil || dropped {
-		return false, nil, err
-	}
-	return true, typ, nil
-}
-
-// TODO (lucy): Should this just take a database name? We're separately
-// resolving the database name in lots of places where we (indirectly) call
-// this.
-func (tc *Collection) getUserDefinedSchemaByName(
-	ctx context.Context, txn *kv.Txn, dbID descpb.ID, schemaName string, flags tree.SchemaLookupFlags,
-) (catalog.SchemaDescriptor, error) {
-	getSchemaByName := func() (found bool, _ catalog.Descriptor, err error) {
-		if descFound, refuseFurtherLookup, desc, err := tc.getSyntheticOrUncommittedDescriptor(
-			dbID, keys.RootNamespaceID, schemaName, flags.RequireMutable,
-		); err != nil || refuseFurtherLookup {
-			return false, nil, err
-		} else if descFound {
-			log.VEventf(ctx, 2, "found uncommitted descriptor %d", desc.GetID())
-			return true, desc, nil
-		}
-
-		if flags.AvoidCached || flags.RequireMutable || lease.TestingTableLeasesAreDisabled() {
-			return tc.kv.getByName(
-				ctx, txn, dbID, keys.RootNamespaceID, schemaName, flags.RequireMutable,
-			)
-		}
-
-		// Look up whether the schema is on the database descriptor and return early
-		// if it's not.
-		_, dbDesc, err := tc.GetImmutableDatabaseByID(
-			ctx, txn, dbID, tree.DatabaseLookupFlags{Required: true},
-		)
-		if err != nil {
-			return false, nil, err
-		}
-		schemaID := dbDesc.GetSchemaID(schemaName)
-		if schemaID == descpb.InvalidID {
-			return false, nil, nil
-		}
-		foundSchemaName := dbDesc.GetNonDroppedSchemaName(schemaID)
-		if foundSchemaName != schemaName {
-			// If there's another schema name entry with the same ID as this one, then
-			// the schema has been renamed, so don't return anything.
-			if foundSchemaName != "" {
-				return false, nil, nil
-			}
-			// Otherwise, the schema has been dropped. Return early, except in the
-			// specific case where flags.Required and flags.IncludeDropped are both
-			// true, which forces us to look up the dropped descriptor and return it.
-			if !flags.Required {
-				return false, nil, nil
-			}
-			if !flags.IncludeDropped {
-				return false, nil, catalog.NewInactiveDescriptorError(catalog.ErrDescriptorDropped)
-			}
-		}
-
-		// If we have a schema ID from the database, get the schema descriptor. Since
-		// the schema and database descriptors are updated in the same transaction,
-		// their leased "versions" (not the descriptor version, but the state in the
-		// abstract sequence of states in adding, renaming, or dropping a schema) can
-		// differ by at most 1 while waiting for old leases to drain. So false
-		// negatives can occur from the database lookup, in some sense, if we have a
-		// lease on the latest version of the schema and on the previous version of
-		// the database which doesn't reflect the changes to the schema. But this
-		// isn't a problem for correctness; it can only happen on other sessions
-		// before the schema change has returned results.
-		desc, err := tc.getDescriptorByID(ctx, txn, schemaID, flags)
-		if err != nil {
-			if errors.Is(err, catalog.ErrDescriptorNotFound) ||
-				errors.Is(err, catalog.ErrDescriptorDropped) {
-				return false, nil, nil
-			}
-			return false, nil, err
-		}
-		return true, desc, nil
-	}
-
-	found, desc, err := getSchemaByName()
-	if err != nil {
-		return nil, err
-	} else if !found {
-		if flags.Required {
-			return nil, sqlerrors.NewUndefinedSchemaError(schemaName)
-		}
-		return nil, nil
-	}
-	schema, ok := desc.(catalog.SchemaDescriptor)
-	if !ok {
-		if flags.Required {
-			return nil, sqlerrors.NewUndefinedSchemaError(schemaName)
-		}
-		return nil, nil
-	}
-	if dropped, err := filterDescriptorState(schema, flags.Required, flags); dropped || err != nil {
-		return nil, err
-	}
-	return schema, nil
-}
-
-// filterDescriptorState wraps the more general catalog function to swallow
-// the error if the descriptor is being dropped and the descriptor is not
-// required. In that case, dropped will be true. A return value of false, nil
-// means this descriptor is okay given the flags.
-// TODO (lucy): We would like the ByID methods to ignore the Required flag and
-// unconditionally return an error for dropped descriptors if IncludeDropped is
-// not set, so we can't just pass the flags passed into the methods into this
-// function, hence the boolean argument. This is the only user of
-// catalog.FilterDescriptorState which needs to pass in nontrivial flags, at
-// time of writing, so we should clean up the interface around this bit of
-// functionality.
-func filterDescriptorState(
-	desc catalog.Descriptor, required bool, flags tree.CommonLookupFlags,
-) (dropped bool, _ error) {
-	flags = tree.CommonLookupFlags{
-		Required:       required,
-		IncludeOffline: flags.IncludeOffline,
-		IncludeDropped: flags.IncludeDropped,
-	}
-	if err := catalog.FilterDescriptorState(desc, flags); err != nil {
-		if required || !errors.Is(err, catalog.ErrDescriptorDropped) {
-			return false, err
-		}
-		return true, nil
-	}
-	return false, nil
-}
-
-// GetMutableSchemaByName resolves the schema and, if applicable, returns a
-// mutable descriptor usable by the transaction. RequireMutable is ignored.
-func (tc *Collection) GetMutableSchemaByName(
-	ctx context.Context, txn *kv.Txn, dbID descpb.ID, schemaName string, flags tree.SchemaLookupFlags,
-) (catalog.SchemaDescriptor, error) {
-	flags.RequireMutable = true
-	return tc.getSchemaByName(ctx, txn, dbID, schemaName, flags)
-}
-
-// GetImmutableSchemaByName resolves the schema and, if applicable, returns an
-// immutable descriptor usable by the transaction. RequireMutable is ignored.
-func (tc *Collection) GetImmutableSchemaByName(
-	ctx context.Context, txn *kv.Txn, dbID descpb.ID, schemaName string, flags tree.SchemaLookupFlags,
-) (catalog.SchemaDescriptor, error) {
-	flags.RequireMutable = false
-	return tc.getSchemaByName(ctx, txn, dbID, schemaName, flags)
-}
-
-// GetSchemaByName returns true and a ResolvedSchema object if the target schema
-// exists under the target database.
-func (tc *Collection) GetSchemaByName(
-	ctx context.Context, txn *kv.Txn, dbID descpb.ID, scName string, flags tree.SchemaLookupFlags,
-) (catalog.SchemaDescriptor, error) {
-	return tc.getSchemaByName(ctx, txn, dbID, scName, flags)
-}
-
-// getSchemaByName resolves the schema and, if applicable, returns a descriptor
-// usable by the transaction.
-func (tc *Collection) getSchemaByName(
-	ctx context.Context, txn *kv.Txn, dbID descpb.ID, schemaName string, flags tree.SchemaLookupFlags,
-) (catalog.SchemaDescriptor, error) {
-	// Fast path public schema, as it is always found.
-	if schemaName == tree.PublicSchema {
-		return schemadesc.GetPublicSchema(), nil
-	}
-
-	if tc.virtualSchemas != nil {
-		if sc, ok := tc.virtualSchemas.GetVirtualSchema(schemaName); ok {
-			return sc.Desc(), nil
-		}
-	}
-
-	// If a temp schema is requested, check if it's for the current session, or
-	// else fall back to reading from the store.
-	if strings.HasPrefix(schemaName, catconstants.PgTempSchemaName) {
-		if tc.sessionData != nil {
-			if schemaName == catconstants.PgTempSchemaName ||
-				schemaName == tc.sessionData.SearchPath.GetTemporarySchemaName() {
-				schemaID, found := tc.sessionData.GetTemporarySchemaIDForDb(uint32(dbID))
-				if found {
-					return schemadesc.NewTemporarySchema(
-						tc.sessionData.SearchPath.GetTemporarySchemaName(),
-						descpb.ID(schemaID),
-						dbID,
-					), nil
-				}
-			}
-		}
-		exists, schemaID, err := catalogkv.ResolveSchemaID(ctx, txn, tc.codec(), dbID, schemaName)
-		if err != nil {
-			return nil, err
-		} else if !exists {
-			if flags.Required {
-				return nil, sqlerrors.NewUndefinedSchemaError(schemaName)
-			}
-			return nil, nil
-		}
-		return schemadesc.NewTemporarySchema(
-			schemaName,
-			schemaID,
-			dbID,
-		), nil
-	}
-
-	// Otherwise, the schema is user-defined. Get the descriptor.
-	desc, err := tc.getUserDefinedSchemaByName(ctx, txn, dbID, schemaName, flags)
-	if err != nil || desc == nil {
-		return nil, err
-	}
-	return desc, nil
-}
-
-// GetMutableDatabaseByID returns a mutable database descriptor with
-// properties according to the provided lookup flags. RequireMutable is ignored.
-func (tc *Collection) GetMutableDatabaseByID(
-	ctx context.Context, txn *kv.Txn, dbID descpb.ID, flags tree.DatabaseLookupFlags,
-) (bool, *dbdesc.Mutable, error) {
-	flags.RequireMutable = true
-	found, desc, err := tc.getDatabaseByID(ctx, txn, dbID, flags)
-	if err != nil || !found {
-		return false, nil, err
-	}
-	return true, desc.(*dbdesc.Mutable), nil
-}
-
-var _ = (*Collection)(nil).GetMutableDatabaseByID
-
-// GetImmutableDatabaseByID returns an immutable database descriptor with
-// properties according to the provided lookup flags. RequireMutable is ignored.
-func (tc *Collection) GetImmutableDatabaseByID(
-	ctx context.Context, txn *kv.Txn, dbID descpb.ID, flags tree.DatabaseLookupFlags,
-) (bool, catalog.DatabaseDescriptor, error) {
-	flags.RequireMutable = false
-	return tc.getDatabaseByID(ctx, txn, dbID, flags)
-}
-
-func (tc *Collection) getDatabaseByID(
-	ctx context.Context, txn *kv.Txn, dbID descpb.ID, flags tree.DatabaseLookupFlags,
-) (bool, catalog.DatabaseDescriptor, error) {
-	desc, err := tc.getDescriptorByID(ctx, txn, dbID, flags)
-	if err != nil {
-		if errors.Is(err, catalog.ErrDescriptorNotFound) {
-			if flags.Required {
-				return false, nil, sqlerrors.NewUndefinedDatabaseError(fmt.Sprintf("[%d]", dbID))
-			}
-			return false, nil, nil
-		}
-		return false, nil, err
-	}
-	db, ok := desc.(catalog.DatabaseDescriptor)
-	if !ok {
-		return false, nil, sqlerrors.NewUndefinedDatabaseError(fmt.Sprintf("[%d]", dbID))
-	}
-	return true, db, nil
-}
-
-// GetMutableTableByID returns a mutable table descriptor with
-// properties according to the provided lookup flags. RequireMutable is ignored.
-// Required is ignored, and an error is always returned if no descriptor with
-// the ID exists.
-func (tc *Collection) GetMutableTableByID(
-	ctx context.Context, txn *kv.Txn, tableID descpb.ID, flags tree.ObjectLookupFlags,
-) (*tabledesc.Mutable, error) {
-	flags.RequireMutable = true
-	desc, err := tc.getTableByID(ctx, txn, tableID, flags)
-	if err != nil {
-		return nil, err
-	}
-	return desc.(*tabledesc.Mutable), nil
-}
-
-// GetImmutableTableByID returns an immutable table descriptor with
-// properties according to the provided lookup flags. RequireMutable is ignored.
-// Required is ignored, and an error is always returned if no descriptor with
-// the ID exists.
-func (tc *Collection) GetImmutableTableByID(
-	ctx context.Context, txn *kv.Txn, tableID descpb.ID, flags tree.ObjectLookupFlags,
-) (catalog.TableDescriptor, error) {
-	flags.RequireMutable = false
-	desc, err := tc.getTableByID(ctx, txn, tableID, flags)
-	if err != nil {
-		return nil, err
-	}
-	return desc, nil
-}
-
-func (tc *Collection) getTableByID(
-	ctx context.Context, txn *kv.Txn, tableID descpb.ID, flags tree.ObjectLookupFlags,
-) (catalog.TableDescriptor, error) {
-	desc, err := tc.getDescriptorByID(ctx, txn, tableID, flags.CommonLookupFlags)
-	if err != nil {
-		if errors.Is(err, catalog.ErrDescriptorNotFound) {
-			return nil, sqlerrors.NewUndefinedRelationError(
-				&tree.TableRef{TableID: int64(tableID)})
-		}
-		return nil, err
-	}
-	table, ok := desc.(catalog.TableDescriptor)
-	if !ok {
-		return nil, sqlerrors.NewUndefinedRelationError(
-			&tree.TableRef{TableID: int64(tableID)})
-	}
-	hydrated, err := tc.hydrateTypesInTableDesc(ctx, txn, table)
-	if err != nil {
-		return nil, err
-	}
-	return hydrated, nil
-}
-
-func (tc *Collection) getDescriptorByID(
-	ctx context.Context, txn *kv.Txn, id descpb.ID, flags tree.CommonLookupFlags,
-) (catalog.Descriptor, error) {
-	return tc.getDescriptorByIDMaybeSetTxnDeadline(
-		ctx, txn, id, flags, false /* setTxnDeadline */)
-}
-
 // SkipValidationOnWrite avoids validating uncommitted descriptors prior to
 // a transaction commit.
 func (tc *Collection) SkipValidationOnWrite() {
 	tc.skipValidationOnWrite = true
-}
-
-// getDescriptorByIDMaybeSetTxnDeadline returns a descriptor according to the
-// provided lookup flags. Note that flags.Required is ignored, and an error is
-// always returned if no descriptor with the ID exists.
-func (tc *Collection) getDescriptorByIDMaybeSetTxnDeadline(
-	ctx context.Context, txn *kv.Txn, id descpb.ID, flags tree.CommonLookupFlags, setTxnDeadline bool,
-) (catalog.Descriptor, error) {
-	getDescriptorByID := func() (catalog.Descriptor, error) {
-		if vd, err := tc.maybeGetVirtualDescriptorByID(
-			ctx, id, flags,
-		); vd != nil || err != nil {
-			return vd, err
-		}
-		if found, sd := tc.synthetic.getByID(id); found {
-			if flags.RequireMutable {
-				return nil, newMutableSyntheticDescriptorAssertionError(sd.GetID())
-			}
-			return sd, nil
-		}
-		if ud := tc.kv.getUncommittedByID(id); ud != nil {
-			log.VEventf(ctx, 2, "found uncommitted descriptor %d", id)
-			if flags.RequireMutable {
-				return ud.mutable, nil
-			}
-			return ud.immutable, nil
-		}
-
-		if flags.AvoidCached || flags.RequireMutable || lease.TestingTableLeasesAreDisabled() {
-			return tc.kv.getByID(ctx, txn, id, flags.RequireMutable)
-		}
-
-		// If we have already read all of the descriptor, use it as a negative
-		// cache to short-circuit a lookup we know will be doomed to fail.
-		//
-		// TODO(ajwerner): More generally leverage this set of kv descriptors on
-		// the resolution path.
-		if tc.kv.idDefinitelyDoesNotExist(id) {
-			return nil, catalog.ErrDescriptorNotFound
-		}
-
-		desc, shouldReadFromStore, err := tc.leased.getByID(ctx, txn, id, setTxnDeadline)
-		if err != nil {
-			return nil, err
-		}
-		if shouldReadFromStore {
-			return tc.kv.getByID(ctx, txn, id, flags.RequireMutable)
-		}
-		return desc, nil
-	}
-
-	desc, err := getDescriptorByID()
-	if err != nil {
-		return nil, err
-	}
-	if dropped, err := filterDescriptorState(desc, true /* required */, flags); err != nil || dropped {
-		// This is a special case for tables in the adding state: Roughly speaking,
-		// we always need to resolve tables in the adding state by ID when they were
-		// newly created in the transaction for DDL statements and for some
-		// information queries (but not for ordinary name resolution for queries/
-		// DML), but we also need to make these tables public in the schema change
-		// job in a separate transaction.
-		// TODO (lucy): We need something like an IncludeAdding flag so that callers
-		// can specify this behavior, instead of having the collection infer the
-		// desired behavior based on the flags (and likely producing unintended
-		// behavior). See the similar comment on etDescriptorByName, which covers
-		// the ordinary name resolution path as well as DDL statements.
-		if desc.Adding() && (desc.IsUncommittedVersion() || flags.AvoidCached || flags.RequireMutable) {
-			return desc, nil
-		}
-		return nil, err
-	}
-	return desc, nil
-}
-
-// GetMutableTableVersionByID is a variant of sqlbase.getTableDescFromID which returns a mutable
-// table descriptor of the table modified in the same transaction.
-// Deprecated in favor of GetMutableTableByID.
-// TODO (lucy): Usages should be replaced with GetMutableTableByID, but this
-// needs a careful look at what flags should be passed in at each call site.
-func (tc *Collection) GetMutableTableVersionByID(
-	ctx context.Context, tableID descpb.ID, txn *kv.Txn,
-) (*tabledesc.Mutable, error) {
-	return tc.GetMutableTableByID(ctx, txn, tableID, tree.ObjectLookupFlags{
-		CommonLookupFlags: tree.CommonLookupFlags{
-			IncludeOffline: true,
-			IncludeDropped: true,
-		},
-	})
-}
-
-// GetMutableDescriptorByID returns a mutable implementation of the descriptor
-// with the requested id. An error is returned if no descriptor exists.
-// Deprecated in favor of GetMutableDescriptorByIDWithFlags.
-func (tc *Collection) GetMutableDescriptorByID(
-	ctx context.Context, id descpb.ID, txn *kv.Txn,
-) (catalog.MutableDescriptor, error) {
-	return tc.GetMutableDescriptorByIDWithFlags(ctx, txn, id, tree.CommonLookupFlags{
-		IncludeOffline: true,
-		IncludeDropped: true,
-	})
-}
-
-// GetMutableDescriptorByIDWithFlags returns a mutable implementation of the
-// descriptor with the requested id. An error is returned if no descriptor exists.
-// TODO (lucy): This is meant to replace GetMutableDescriptorByID. Once it does,
-// rename this function.
-func (tc *Collection) GetMutableDescriptorByIDWithFlags(
-	ctx context.Context, txn *kv.Txn, id descpb.ID, flags tree.CommonLookupFlags,
-) (catalog.MutableDescriptor, error) {
-	log.VEventf(ctx, 2, "planner getting mutable descriptor for id %d", id)
-	flags.RequireMutable = true
-	desc, err := tc.getDescriptorByID(ctx, txn, id, flags)
-	if err != nil {
-		return nil, err
-	}
-	return desc.(catalog.MutableDescriptor), nil
-}
-
-// GetImmutableDescriptorByID returns an immmutable implementation of the
-// descriptor with the requested id. An error is returned if no descriptor exists.
-// Deprecated in favor of GetMutableDescriptorByIDWithFlags.
-func (tc *Collection) GetImmutableDescriptorByID(
-	ctx context.Context, txn *kv.Txn, id descpb.ID, flags tree.CommonLookupFlags,
-) (catalog.Descriptor, error) {
-	log.VEventf(ctx, 2, "planner getting immutable descriptor for id %d", id)
-	flags.RequireMutable = false
-	return tc.getDescriptorByID(ctx, txn, id, flags)
-}
-
-// GetMutableSchemaByID returns a ResolvedSchema wrapping a mutable
-// descriptor, if applicable. RequireMutable is ignored.
-// Required is ignored, and an error is always returned if no descriptor with
-// the ID exists.
-func (tc *Collection) GetMutableSchemaByID(
-	ctx context.Context, txn *kv.Txn, schemaID descpb.ID, flags tree.SchemaLookupFlags,
-) (catalog.SchemaDescriptor, error) {
-	flags.RequireMutable = true
-	return tc.getSchemaByID(ctx, txn, schemaID, flags)
-}
-
-var _ = (*Collection)(nil).GetMutableSchemaByID
-
-// GetImmutableSchemaByID returns a ResolvedSchema wrapping an immutable
-// descriptor, if applicable. RequireMutable is ignored.
-// Required is ignored, and an error is always returned if no descriptor with
-// the ID exists.
-func (tc *Collection) GetImmutableSchemaByID(
-	ctx context.Context, txn *kv.Txn, schemaID descpb.ID, flags tree.SchemaLookupFlags,
-) (catalog.SchemaDescriptor, error) {
-	flags.RequireMutable = false
-	return tc.getSchemaByID(ctx, txn, schemaID, flags)
-}
-
-func (tc *Collection) getSchemaByID(
-	ctx context.Context, txn *kv.Txn, schemaID descpb.ID, flags tree.SchemaLookupFlags,
-) (catalog.SchemaDescriptor, error) {
-	if schemaID == keys.PublicSchemaID {
-		return schemadesc.GetPublicSchema(), nil
-	}
-
-	// We have already considered if the schemaID is PublicSchemaID,
-	// if the id appears in staticSchemaIDMap, it must map to a virtual schema.
-	if tc.virtualSchemas != nil {
-		if sc, ok := tc.virtualSchemas.GetVirtualSchemaByID(schemaID); ok {
-			return sc.Desc(), nil
-		}
-	}
-
-	// If this collection is attached to a session and the session has created
-	// a temporary schema, then check if the schema ID matches.
-	if tc.sessionData != nil {
-		dbID, exists := tc.sessionData.MaybeGetDatabaseForTemporarySchemaID(uint32(schemaID))
-		if exists {
-			return schemadesc.NewTemporarySchema(
-				tc.sessionData.SearchPath.GetTemporarySchemaName(),
-				schemaID,
-				descpb.ID(dbID),
-			), nil
-		}
-	}
-
-	// Otherwise, fall back to looking up the descriptor with the desired ID.
-	desc, err := tc.getDescriptorByID(ctx, txn, schemaID, flags)
-	if err != nil {
-		return nil, err
-	}
-
-	schemaDesc, ok := desc.(catalog.SchemaDescriptor)
-	if !ok {
-		return nil, pgerror.Newf(pgcode.WrongObjectType,
-			"descriptor %d was not a schema", schemaID)
-	}
-
-	return schemaDesc, nil
-}
-
-// hydrateTypesInTableDesc installs user defined type metadata in all types.T
-// present in the input TableDescriptor. It always returns the same type of
-// TableDescriptor that was passed in. It ensures that ImmutableTableDescriptors
-// are not modified during the process of metadata installation. Dropped tables
-// do not get hydrated.
-//
-// TODO(ajwerner): This should accept flags to indicate whether we can resolve
-// offline descriptors.
-func (tc *Collection) hydrateTypesInTableDesc(
-	ctx context.Context, txn *kv.Txn, desc catalog.TableDescriptor,
-) (catalog.TableDescriptor, error) {
-	if desc.Dropped() {
-		return desc, nil
-	}
-	switch t := desc.(type) {
-	case *tabledesc.Mutable:
-		// It is safe to hydrate directly into Mutable since it is
-		// not shared. When hydrating mutable descriptors, use the mutable access
-		// method to access types.
-		getType := func(ctx context.Context, id descpb.ID) (tree.TypeName, catalog.TypeDescriptor, error) {
-			desc, err := tc.GetMutableTypeVersionByID(ctx, txn, id)
-			if err != nil {
-				return tree.TypeName{}, nil, err
-			}
-			dbDesc, err := tc.GetMutableDescriptorByID(ctx, desc.ParentID, txn)
-			if err != nil {
-				return tree.TypeName{}, nil, err
-			}
-			sc, err := tc.getSchemaByID(
-				ctx, txn, desc.ParentSchemaID,
-				tree.SchemaLookupFlags{
-					IncludeOffline: true,
-					RequireMutable: true,
-				},
-			)
-			if err != nil {
-				return tree.TypeName{}, nil, err
-			}
-			name := tree.MakeNewQualifiedTypeName(dbDesc.GetName(), sc.GetName(), desc.Name)
-			return name, desc, nil
-		}
-
-		return desc, typedesc.HydrateTypesInTableDescriptor(ctx, t.TableDesc(), typedesc.TypeLookupFunc(getType))
-	case catalog.TableDescriptor:
-		// ImmutableTableDescriptors need to be copied before hydration, because
-		// they are potentially read by multiple threads. If there aren't any user
-		// defined types in the descriptor, then return early.
-		if !t.ContainsUserDefinedTypes() {
-			return desc, nil
-		}
-
-		getType := typedesc.TypeLookupFunc(func(
-			ctx context.Context, id descpb.ID,
-		) (tree.TypeName, catalog.TypeDescriptor, error) {
-			desc, err := tc.GetImmutableTypeByID(ctx, txn, id, tree.ObjectLookupFlags{})
-			if err != nil {
-				return tree.TypeName{}, nil, err
-			}
-			_, dbDesc, err := tc.GetImmutableDatabaseByID(ctx, txn, desc.GetParentID(),
-				tree.DatabaseLookupFlags{Required: true})
-			if err != nil {
-				return tree.TypeName{}, nil, err
-			}
-			sc, err := tc.GetImmutableSchemaByID(
-				ctx, txn, desc.GetParentSchemaID(), tree.SchemaLookupFlags{})
-			if err != nil {
-				return tree.TypeName{}, nil, err
-			}
-			name := tree.MakeNewQualifiedTypeName(dbDesc.GetName(), sc.GetName(), desc.GetName())
-			return name, desc, nil
-		})
-
-		// Utilize the cache of hydrated tables if we have one.
-		if tc.hydratedTables != nil {
-			hydrated, err := tc.hydratedTables.GetHydratedTableDescriptor(ctx, t, getType)
-			if err != nil {
-				return nil, err
-			}
-			if hydrated != nil {
-				return hydrated, nil
-			}
-			// The cache decided not to give back a hydrated descriptor, likely
-			// because either we've modified the table or one of the types or because
-			// this transaction has a stale view of one of the relevant descriptors.
-			// Proceed to hydrating a fresh copy.
-		}
-
-		// Make a copy of the underlying descriptor before hydration.
-		descBase := protoutil.Clone(t.TableDesc()).(*descpb.TableDescriptor)
-		if err := typedesc.HydrateTypesInTableDescriptor(ctx, descBase, getType); err != nil {
-			return nil, err
-		}
-		if t.IsUncommittedVersion() {
-			return tabledesc.NewBuilderForUncommittedVersion(descBase).BuildImmutableTable(), nil
-		}
-		return tabledesc.NewBuilder(descBase).BuildImmutableTable(), nil
-	default:
-		return desc, nil
-	}
 }
 
 // ReleaseSpecifiedLeases releases the leases for the descriptors with ids in
@@ -1249,69 +289,6 @@ func (tc *Collection) ValidateUncommittedDescriptors(ctx context.Context, txn *k
 	return tc.kv.validateUncommittedDescriptors(ctx, txn)
 }
 
-// User defined type accessors.
-
-// GetMutableTypeVersionByID is the equivalent of GetMutableTableDescriptorByID
-// but for accessing types.
-// Deprecated in favor of GetMutableTypeByID.
-// TODO (lucy): Usages should be replaced with GetMutableTypeByID, but this
-// needs a careful look at what flags should be passed in at each call site.
-func (tc *Collection) GetMutableTypeVersionByID(
-	ctx context.Context, txn *kv.Txn, typeID descpb.ID,
-) (*typedesc.Mutable, error) {
-	return tc.GetMutableTypeByID(ctx, txn, typeID, tree.ObjectLookupFlags{
-		CommonLookupFlags: tree.CommonLookupFlags{
-			IncludeOffline: true,
-			IncludeDropped: true,
-		},
-	})
-}
-
-// GetMutableTypeByID returns a mutable type descriptor with
-// properties according to the provided lookup flags. RequireMutable is ignored.
-// Required is ignored, and an error is always returned if no descriptor with
-// the ID exists.
-func (tc *Collection) GetMutableTypeByID(
-	ctx context.Context, txn *kv.Txn, typeID descpb.ID, flags tree.ObjectLookupFlags,
-) (*typedesc.Mutable, error) {
-	flags.RequireMutable = true
-	desc, err := tc.getTypeByID(ctx, txn, typeID, flags)
-	if err != nil {
-		return nil, err
-	}
-	return desc.(*typedesc.Mutable), nil
-}
-
-// GetImmutableTypeByID returns an immutable type descriptor with
-// properties according to the provided lookup flags. RequireMutable is ignored.
-// Required is ignored, and an error is always returned if no descriptor with
-// the ID exists.
-func (tc *Collection) GetImmutableTypeByID(
-	ctx context.Context, txn *kv.Txn, typeID descpb.ID, flags tree.ObjectLookupFlags,
-) (catalog.TypeDescriptor, error) {
-	flags.RequireMutable = false
-	return tc.getTypeByID(ctx, txn, typeID, flags)
-}
-
-func (tc *Collection) getTypeByID(
-	ctx context.Context, txn *kv.Txn, typeID descpb.ID, flags tree.ObjectLookupFlags,
-) (catalog.TypeDescriptor, error) {
-	desc, err := tc.getDescriptorByID(ctx, txn, typeID, flags.CommonLookupFlags)
-	if err != nil {
-		if errors.Is(err, catalog.ErrDescriptorNotFound) {
-			return nil, pgerror.Newf(
-				pgcode.UndefinedObject, "type with ID %d does not exist", typeID)
-		}
-		return nil, err
-	}
-	typ, ok := desc.(catalog.TypeDescriptor)
-	if !ok {
-		return nil, pgerror.Newf(
-			pgcode.UndefinedObject, "type with ID %d does not exist", typeID)
-	}
-	return typ, nil
-}
-
 // getSyntheticOrUncommittedDescriptor attempts to look up a descriptor in the
 // set of synthetic descriptors, followed by the set of uncommitted descriptors.
 func (tc *Collection) getSyntheticOrUncommittedDescriptor(
@@ -1339,16 +316,6 @@ func newMutableSyntheticDescriptorAssertionError(id descpb.ID) error {
 	return errors.AssertionFailedf("attempted mutable access of synthetic descriptor %d", id)
 }
 
-// GetUncommittedTableByID returns an uncommitted table by its ID.
-func (tc *Collection) GetUncommittedTableByID(id descpb.ID) *tabledesc.Mutable {
-	if ud := tc.kv.getUncommittedByID(id); ud != nil {
-		if table, ok := ud.mutable.(*tabledesc.Mutable); ok {
-			return table
-		}
-	}
-	return nil
-}
-
 // GetAllDescriptors returns all descriptors visible by the transaction,
 // first checking the Collection's cached descriptors for validity if validate
 // is set to true before defaulting to a key-value scan, if necessary.
@@ -1356,78 +323,6 @@ func (tc *Collection) GetAllDescriptors(
 	ctx context.Context, txn *kv.Txn,
 ) ([]catalog.Descriptor, error) {
 	return tc.kv.getAllDescriptors(ctx, txn)
-}
-
-// HydrateGivenDescriptors installs type metadata in the types present for all
-// table descriptors in the slice of descriptors. It is exported so resolution
-// on sets of descriptors can hydrate a set of descriptors (i.e. on BACKUPs).
-func HydrateGivenDescriptors(ctx context.Context, descs []catalog.Descriptor) error {
-	// Collect the needed information to set up metadata in those types.
-	dbDescs := make(map[descpb.ID]catalog.DatabaseDescriptor)
-	typDescs := make(map[descpb.ID]catalog.TypeDescriptor)
-	schemaDescs := make(map[descpb.ID]catalog.SchemaDescriptor)
-	for _, desc := range descs {
-		switch desc := desc.(type) {
-		case catalog.DatabaseDescriptor:
-			dbDescs[desc.GetID()] = desc
-		case catalog.TypeDescriptor:
-			typDescs[desc.GetID()] = desc
-		case catalog.SchemaDescriptor:
-			schemaDescs[desc.GetID()] = desc
-		}
-	}
-	// If we found any type descriptors, that means that some of the tables we
-	// scanned might have types that need hydrating.
-	if len(typDescs) > 0 {
-		// Since we just scanned all the descriptors, we already have everything
-		// we need to hydrate our types. Set up an accessor for the type hydration
-		// method to look into the scanned set of descriptors.
-		typeLookup := func(ctx context.Context, id descpb.ID) (tree.TypeName, catalog.TypeDescriptor, error) {
-			typDesc, ok := typDescs[id]
-			if !ok {
-				n := tree.MakeUnresolvedName(fmt.Sprintf("[%d]", id))
-				return tree.TypeName{}, nil, sqlerrors.NewUndefinedObjectError(&n,
-					tree.TypeObject)
-			}
-			dbDesc, ok := dbDescs[typDesc.GetParentID()]
-			if !ok {
-				n := fmt.Sprintf("[%d]", typDesc.GetParentID())
-				return tree.TypeName{}, nil, sqlerrors.NewUndefinedDatabaseError(n)
-			}
-			// We don't use the collection's ResolveSchemaByID method here because
-			// we already have all of the descriptors. User defined types are only
-			// members of the public schema or a user defined schema, so those are
-			// the only cases we have to consider here.
-			var scName string
-			switch typDesc.GetParentSchemaID() {
-			case keys.PublicSchemaID:
-				scName = tree.PublicSchema
-			default:
-				scName = schemaDescs[typDesc.GetParentSchemaID()].GetName()
-			}
-			name := tree.MakeNewQualifiedTypeName(dbDesc.GetName(), scName, typDesc.GetName())
-			return name, typDesc, nil
-		}
-		// Now hydrate all table descriptors.
-		for i := range descs {
-			desc := descs[i]
-			// Never hydrate dropped descriptors.
-			if desc.Dropped() {
-				continue
-			}
-			tblDesc, ok := desc.(catalog.TableDescriptor)
-			if ok {
-				if err := typedesc.HydrateTypesInTableDescriptor(
-					ctx,
-					tblDesc.TableDesc(),
-					typedesc.TypeLookupFunc(typeLookup),
-				); err != nil {
-					return err
-				}
-			}
-		}
-	}
-	return nil
 }
 
 // GetAllDatabaseDescriptors returns all database descriptors visible by the
